categorizes all the constants
