+++ src/js/constants/classnames.ts
@@ -1,0 +1,2 @@
+
+export const SHABAD_CONTENT_CLASSNAME = 'shabad-content';
