+++ src/js/constants/colors.ts
@@ -1,0 +1,3 @@
+export const LARIVAAR_ASSIST_COLOR = '#f39c1d';
+export const HIGHLIGHTED_SEARCH_COLOR = 'rgb(1, 102, 155)';
+export const NORMAL_SEARCH_COLOR = 'rgba(1, 102, 155, 0.3)';
