+++ src/js/constants/default-option-values.ts
@@ -1,0 +1,24 @@
+import { SEARCH_TYPES } from "./search-types";
+
+export const DEFAULT_SEARCH_TYPE = SEARCH_TYPES.FIRST_LETTERS;
+export const DEFAULT_SEARCH_SOURCE = 'all';
+export const DEFAULT_TRANSLATION_LANGUAGES = ['english'];
+export const DEFAULT_TRANSLITERATION_LANGUAGES = ['english'];
+export const DEFAULT_LARIVAAR = false;
+export const DEFAULT_LARIVAAR_ASSIST = false;
+export const DEFAULT_UNICODE = false;
+export const DEFAULT_SPLIT_VIEW = false;
+export const DEFAULT_FONT_SIZE = 2;
+export const DEFAULT_TRANSLATION_FONT_SIZE = 1.2;
+export const DEFAULT_TRANSLITERATION_FONT_SIZE = 1.2;
+export const DEFAULT_LINE_HEIGHT = 1.2;
+export const DEFAULT_FONT_FAMILY = 'gurmukhi_heavy';
+export const DEFAULT_PAGE_TITLE = 'SikhiToTheMax';
+export const DEFAULT_DARK_MODE = false;
+export const DEFAULT_AUTO_SCROLL_MODE = false;
+export const DEFAULT_IS_AUTOSCROLLING = false;
+export const DEFAULT_PARAGRAPH_MODE = false;
+export const DEFAULT_CENTER_ALIGN_GURBANI = false;
+export const DEFAULT_VISRAAMS = false;
+export const DEFAULT_VISRAAM_SOURCE = 'sttm';
+export const DEFAULT_VISRAAM_STYLE = 'colored-words';
