+++ src/js/constants/font-options.ts
@@ -1,0 +1,11 @@
+export const FONT_OPTIONS = {
+  anmol_lipi: 'Anmol Lipi',
+  anmol_bold: 'Anmol Lipi Bold',
+  notosans: 'Noto Sans',
+  notosans_bold: 'Noto Sans Bold',
+  gurmukhi_normal: 'Gurbani Akhar',
+  gurmukhi_heavy: 'Gurbani Akhar Heavy',
+  gurmukhi_thick: 'Gurbani Akhar Thick',
+  prabhki: 'Prabhki',
+  unicode_font: 'Unicode',
+};
