+++ src/js/constants/languages-supported.ts
@@ -1,0 +1,2 @@
+export const TRANSLATION_LANGUAGES = ['punjabi', 'english', 'spanish'];
+export const TRANSLITERATION_LANGUAGES = ['english', 'hindi', 'shahmukhi'];
