+++ src/js/constants/local-storage-keys.ts
@@ -1,0 +1,26 @@
+export const LOCAL_STORAGE_KEY_FOR_GDPR_NOTICE = 'gdpr_notice';
+export const LOCAL_STORAGE_KEY_FOR_SEARCH_SOURCE = 'searchSource';
+export const LOCAL_STORAGE_KEY_FOR_SEARCH_TYPE = 'searchType';
+export const LOCAL_STORAGE_KEY_FOR_TRANSLATION_LANGUAGES =
+  'translationLanguages';
+export const LOCAL_STORAGE_KEY_FOR_TRANSLITERATION_LANGUAGES =
+  'transliterationLanguages';
+export const LOCAL_STORAGE_KEY_FOR_LARIVAAR_ASSIST = 'larivaarAssist';
+export const LOCAL_STORAGE_KEY_FOR_LARIVAAR = 'larivaar';
+export const LOCAL_STORAGE_KEY_FOR_UNICODE = 'unicode';
+export const LOCAL_STORAGE_KEY_FOR_SPLIT_VIEW = 'splitView';
+export const LOCAL_STORAGE_KEY_FOR_FONT_SIZE = 'fontSize';
+export const LOCAL_STORAGE_KEY_FOR_TRANSLATION_FONT_SIZE = 'translationFontSize';
+export const LOCAL_STORAGE_KEY_FOR_TRANSLITERATION_FONT_SIZE = 'transliterationFontSize';
+export const LOCAL_STORAGE_KEY_FOR_LINE_HEIGHT = 'lineHeight';
+export const LOCAL_STORAGE_KEY_FOR_FONT_FAMILY = 'fontFamily';
+export const LOCAL_STORAGE_KEY_FOR_DARK_MODE = 'darkMode';
+export const LOCAL_STORAGE_KEY_FOR_AUTO_SCROLL_MODE = 'autoScrollMode';
+export const LOCAL_STORAGE_KEY_FOR_VISRAAMS = 'visraams';
+export const LOCAL_STORAGE_KEY_FOR_VISRAAM_SOURCE = 'visraamSource';
+export const LOCAL_STORAGE_KEY_FOR_VISRAAMS_STYLE = 'visraamStyle';
+export const LOCAL_STORAGE_KEY_FOR_PREVIOUSLY_READ_ANG = 'previouslyReadAng';
+export const LOCAL_STORAGE_KEY_FOR_RELEASE = 'release';
+export const LOCAL_STORAGE_KEY_FOR_CENTER_ALIGN_VIEW = 'centerAlignView';
+export const LOCAL_STORAGE_KEY_FOR_SYNC_CODE = 'previous-sync';
+export const LOCAL_STORAGE_KEY_FOR_PARAGRAPH_MODE = 'paragraphMode';
