+++ src/js/constants/misc.ts
@@ -1,0 +1,23 @@
+import { SOURCES, SOURCES_WITH_ANG, TYPES as _TYPES } from '@sttm/banidb';
+import { SEARCH_TYPES } from './search-types';
+
+export { SOURCES, SOURCES_WITH_ANG };
+
+export const BANI_LENGTH_COLS = {
+  short: 'existsSGPC',
+  medium: 'existsMedium',
+  long: 'existsTaksal',
+  extralong: 'existsBuddhaDal',
+};
+
+export const TYPES = _TYPES.filter((value, index) =>
+  Object.values(SEARCH_TYPES).includes(index));
+
+export const SHORT_DOMAIN = 'sttm.co';
+
+export const MAX_ANGS = {
+  G: 1430,
+  D: 1428,
+  B: 40,
+  S: 28
+}
