+++ src/js/constants/page-name.ts
@@ -1,0 +1,8 @@
+export const PAGE_NAME = {
+  ANG: {
+    gurmukhi: 'AMg', unicode: 'ਅੰਗ'
+  },
+  PANNA: {
+    gurmukhi: 'pMnw', unicode: 'ਪੰਨਾ'
+  }
+}
