+++ src/js/constants/placeholders.ts
@@ -1,0 +1,10 @@
+import { SEARCH_TYPES } from "./search-types";
+
+export const PLACEHOLDERS = {
+  [SEARCH_TYPES.FIRST_LETTERS]: ['jmTAq'], // first letters
+  [SEARCH_TYPES.FIRST_LETTERS_ANYWHERE]: ['mqjbe'], // first letter anywhere
+  [SEARCH_TYPES.GURMUKHI_WORD]: ['jo mwgih Twkur Apuny qy'], // gurmukhi
+  [SEARCH_TYPES.ENGLISH_WORD]: ['He has extended His power', true], // translation
+  [SEARCH_TYPES.ROMANIZED]: ['jo mange thakur apne te soi', true], // romanized
+  [SEARCH_TYPES.ANG]: ['123', true], // ang
+};
