+++ src/js/constants/search-types.ts
@@ -1,0 +1,8 @@
+export const SEARCH_TYPES = {
+  FIRST_LETTERS: 0,
+  FIRST_LETTERS_ANYWHERE: 1,
+  GURMUKHI_WORD: 2,
+  ENGLISH_WORD: 3,
+  ROMANIZED: 4,
+  ANG: 5,
+};
