+++ src/js/constants/sync-types.ts
@@ -1,0 +1,6 @@
+
+export const SYNC_TYPES = {
+  SHABAD: 'shabad',
+  CEREMONY: 'ceremony',
+  BANI: 'bani',
+};
