+++ src/js/settings.ts
@@ -2,7 +2,7 @@ import {
   TRANSLATION_LANGUAGES,
   TRANSLITERATION_LANGUAGES,
   FONT_OPTIONS,
-  VISRAAM_CONSTANTS,
+  VISRAAM,
 } from './constants';
 
 import { toggleItemInArray, toFixedFloat } from './util';
@@ -322,17 +322,17 @@ export const ADVANCED_SETTINGS = ({
     },
     {
       type: 'dropdown',
-      label: 'Visraam Source',
+      label: 'VISRAAM Source',
       value: visraamSource,
       action: setVisraamSource,
-      options: VISRAAM_CONSTANTS.SOURCES,
+      options: VISRAAM.SOURCES,
     },
     {
       type: 'dropdown',
-      label: 'Visraam Style',
+      label: 'VISRAAM Style',
       value: visraamStyle,
       action: setVisraamStyle,
-      options: VISRAAM_CONSTANTS.TYPES,
+      options: VISRAAM.TYPES,
     },
     {
       type: 'dropdown',
