+++ src/scss/_vishraam.scss
@@ -1,20 +1,20 @@
 /* stylelint-disable at-rule-no-unknown */
 
-@mixin vishraamStyles($visraam, $yamki) {
-  $visraam-main: $visraam;
-  $visraam-yamki: $yamki;
-  $visraam-sources: ('sttm', 'sttm2', 'igurbani');
+@mixin vishraamStyles($VISRAAM, $yamki) {
+  $VISRAAM-main: $VISRAAM;
+  $VISRAAM-yamki: $yamki;
+  $VISRAAM-sources: ('sttm', 'sttm2', 'igurbani');
 
-  @each $source in $visraam-sources {
+  @each $source in $VISRAAM-sources {
     &.display-visraams.vishraam-vishraam-options-colored-words.vishraam-vishraam-source-#{$source} {
       .gurlipi,
       .unicode {
-        .visraam-#{$source}-main {
-          color: $visraam-main;
+        .VISRAAM-#{$source}-main {
+          color: $VISRAAM-main;
         }
 
-        .visraam-#{$source}-yamki {
-          color: $visraam-yamki;
+        .VISRAAM-#{$source}-yamki {
+          color: $VISRAAM-yamki;
         }
       }
     }
@@ -26,26 +26,26 @@
           border-radius: 15px;
         }
 
-        .padchhed .visraam-#{$source}-main,
-        .padchhed .visraam-#{$source}-yamki {
+        .padchhed .VISRAAM-#{$source}-main,
+        .padchhed .VISRAAM-#{$source}-yamki {
           padding-right: 5px;
         }
 
-        .visraam-#{$source}-main {
+        .VISRAAM-#{$source}-main {
           background:
             linear-gradient(
               to right,
               rgba(229, 229, 229, 0) 20%,
-              rgba($visraam-main, 0.7) 100%
+              rgba($VISRAAM-main, 0.7) 100%
             );
         }
 
-        .visraam-#{$source}-yamki {
+        .VISRAAM-#{$source}-yamki {
           background:
             linear-gradient(
               to right,
               rgba(229, 229, 229, 0) 20%,
-              rgba($visraam-yamki, 0.7) 100%
+              rgba($VISRAAM-yamki, 0.7) 100%
             );
         }
       }
